♻️ refactor : 댓글 불러오기 함수를 커스텀 훅으로 분리 적용
